+++ src/wasm-lib/kcl/src/lib.rs
@@ -30,6 +30,7 @@ pub mod std;
 pub mod test_server;
 pub mod thread;
 pub mod token;
+mod unparser;
 pub mod walk;
 #[cfg(target_arch = "wasm32")]
 pub mod wasm;
